refactor: inline single-caller helpers (obsidian-api, deep-research, output, notebooklm-citations)
- ConnectionTestResult interface → inline return type on testConnection
- Citation factory functions → module-scoped const RegExp
- executeOutput dispatcher → inline switch in handleMultiOutput.map()
- notebooklm-citations.ts module → inline into notebooklm.ts (sole caller);
  delete dedicated unit test (behavior covered by notebooklm.test.ts:167-308
  via public NotebookLMExtractor API)
