+++ src/lib/obsidian-api.ts
@@ -57,18 +57,6 @@ function createTimeoutSignal(ms: number): AbortSignal {
   return controller.signal;
 }
 
-/**
- * Connection test result with detailed status
- */
-interface ConnectionTestResult {
-  /** Server is reachable */
-  reachable: boolean;
-  /** API Key is valid (authentication succeeded) */
-  authenticated: boolean;
-  /** Error message (when failed) */
-  error?: string;
-}
-
 export class ObsidianApiError extends Error {
   readonly status: number;
   constructor(status: number, message: string) {
@@ -130,7 +118,14 @@ export class ObsidianApiClient {
    * Uses /vault/ endpoint which requires authentication.
    * This ensures the API key is validated, not just server reachability.
    */
-  async testConnection(): Promise<ConnectionTestResult> {
+  async testConnection(): Promise<{
+    /** Server is reachable */
+    reachable: boolean;
+    /** API Key is valid (authentication succeeded) */
+    authenticated: boolean;
+    /** Error message (when failed) */
+    error?: string;
+  }> {
     try {
       const response = await fetch(`${this.baseUrl}/vault/`, {
         method: 'GET',
